refactor(generic): use new transport crate
refactor(generic): impl functionality as traits
refactor(generic): move common mod in lib
