+++ ledger-zondax-generic/Cargo.toml
@@ -10,16 +10,10 @@ readme = "README.md"
 categories = ["authentication", "cryptography"]
 keywords = ["ledger", "nano", "blue", "apdu"]
 edition = "2018"
-autobenches = false
-
-[badges]
-circle-ci = { repository = "zondax/ledger-rs" }
-
-[lib]
-name = "ledger_zondax_generic"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
-ledger-apdu = { path = "../ledger-apdu", version = "0.8.0" }
-ledger-transport = { path = "../ledger-transport", version = "0.8.0" }
 thiserror = "1.0"
+
+ledger-transport = "0.8.0"
+async-trait = "0.1.53"
