+++ ledger-zondax-generic/README.md
@@ -2,4 +2,4 @@
 
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 
-Generic library with APDU commands typically used in Ledger apps developed by Zondax
+Common APDU commands typically used in Ledger Apps developed by Zondax
